# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Upload:
    properties:
      createdAt:
        type: string
      file:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    required:
    - file
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a swagger documentation for Costumer APP.
  title: Swagger Costumer APP
  version: "2.0"
paths:
...
  /images:
    post:
      consumes:
      - multipart/form-data
      description: Upload a new asset file and save its metadata to the database
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Asset name
        in: formData
        name: name
        required: true
        type: string
      - description: Asset file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Upload'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new asset
      tags:
      - assets
  /images/{id}:
    delete:
      description: Delete an asset file and its metadata from the database
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an asset
      tags:
      - assets
...
schemes:
- http
- https
swagger: "2.0"